feat: Implement Favorite functionality with controller and router
- Added FavoriteController to handle viewing and creating favorite books.
- Created FavoriteRouter for defining routes related to favorites.
- Integrated favorite routes into the main server file.
- Updated LikesMessageController and LoansController for improved error handling and response structure.
- Enhanced PostMessageController with better query structure and error handling.
- Introduced socket.io for real-time comment notifications.
- Added email notification functionality for loan confirmations.

+++ BackEnd/src/controllers/PostMessageController.js
@@ -1,37 +1,69 @@
-const connection = require("../config/db")
+const connection = require("../config/db");
 
 exports.viewPost = (req, res) => {
-  connection.query(`SELECT * 
-        FROM Post p
-        JOIN Topic t on p.Topic_idTopic = t.idTopic
-        ORDER BY p.created_at DESC
-    `, (err, result) => {
+  const query = `
+    SELECT
+      p.idPost AS id,
+      p.title,
+      p.content,
+      p.image AS image_url,
+      p.created_at,
+      p.updated_at,
+      u.idUser AS user_id,
+      u.username,
+      u.profile_picture_url,
+      t.idTopic,
+      t.title AS topic_title,
+      t.description AS topic_description,
+      (SELECT COUNT(*) FROM likes l WHERE l.post_id = p.idPost) AS likes_count,
+      (SELECT COUNT(*) FROM comments c WHERE c.post_id = p.idPost) AS comments_count
+    FROM Post p
+    JOIN User u ON p.User_idUser = u.idUser
+    JOIN Topic t ON p.Topic_idTopic = t.idTopic
+    ORDER BY p.created_at DESC
+  `;
+
+  connection.query(query, (err, results) => {
     if (err) {
       return res.status(500).json({
         message: "Erro ao se conectar com o servidor.",
         success: false,
         data: err,
-      })
-    } else {
-      return res.status(200).json({
-        message: "Sucesso ao exibir as mensagens.",
-        success: true,
-        data: result,
-      })
+      });
     }
-  })
-}
+    return res.status(200).json({
+      message: "Sucesso ao exibir as mensagens.",
+      success: true,
+      data: results,
+    });
+  });
+};
 // Aqui eu faço diferente das demais, pois, futuramente eu posso querer exibir um histórico pedidos de reserva e para eu mostrar para o usuário eu tenho que fazer uma ligação de todas as tabelas responsáveis por isso.
 
 exports.viewPostByUser = (req, res) => {
-  const User_idUser = req.data.id
+  const User_idUser = req.data.id;
 
   connection.query(
-    `SELECT * 
-        FROM Post p
-        JOIN User u on p.User_idUser = u.idUser
-        WHERE u.idUser = ? 
-        
+    `SELECT
+      p.idPost AS id,
+      p.title,
+      p.content,
+      p.image AS image_url,
+      p.created_at,
+      p.updated_at,
+      u.idUser AS user_id,
+      u.username,
+      u.profile_picture_url,
+      t.idTopic,
+      t.title as topic_title,
+      t.description As topic_description,
+      (SELECT COUNT(*) FROM likes l WHERE l.post_id = p.idPost) AS likes_count,
+      (SELECT COUNT(*) FROM comments c WHERE c.post_id = p.idPost) AS comments_count
+    FROM Post p
+    JOIN User u ON p.User_idUser = u.idUser
+    JOIN Topic t on p.Topic_idTopic = t.idTopic
+    where u.idUser = ?
+    ORDER BY p.created_at DESC
         `,
     [User_idUser],
     (err, result) => {
@@ -40,133 +72,149 @@ exports.viewPostByUser = (req, res) => {
           message: "Erro ao se conectar com o servidor.",
           success: false,
           data: err,
-        })
+        });
       }
       if (result.length === 0) {
         return response.status(400).json({
           success: false,
           message: `Nenhuma publicação encontrada para este usuário com o ID especificado.`,
-        })
-      } else {
-        return res.status(200).json({
-          message: "Publicação encontrada com sucesso.",
-          success: true,
-          data: result,
-        })
+        });
       }
+      return res.status(200).json({
+        message: "Publicação encontrada com sucesso.",
+        success: true,
+        data: result,
+      });
     }
-  )
-}
+  );
+};
 
 exports.createPost = (req, res) => {
-  const { content, Topic_idTopic, User_idUser } = req.body
-  const image = req.file ? req.file.filename : null
+  const { content, Topic_idTopic } = req.body;
+  const image = req.file ? req.file.filename : null;
+  const User_idUser = req.data.id
 
   if (!content || !image || !Topic_idTopic || !User_idUser) {
     return res.status(400).json({
       success: false,
       message: "Preencha todos os campos de cadastro",
-      data: console.log(content, image, Topic_idTopic, User_idUser)
-    })
+      data: console.log(content, image, Topic_idTopic, User_idUser),
+    });
   }
-  
 
-  // verificar se a postagem já está publicada
+  // verificar se o topic existe
   connection.query(
-    "SELECT * FROM Post where content = ? AND User_idUser = ?",
-    [content, User_idUser],
-    (err, result) => {
-      if (err) {
+    "SELECT idTopic, title, image, User_idUser, created_at FROM Topic where idTopic = ?",
+    [Topic_idTopic],
+    (errTopic, resultTopic) => {
+      if (errTopic) {
         return res.status(500).json({
-          message: "Erro ao verificar reservas realizados.",
+          message: "Erro ao verificar o tópico.",
           success: false,
-          data: err,
-        })
+          data: errTopic,
+        });
       }
 
-      // Verifica se já existe uma postagem
-
-      if (result.length > 0) {
+      if (resultTopic.length === 0) {
         return res.status(400).json({
-          message: "Já existe uma postagem igual. Por favor, tente novamente.",
+          message: "Tópico informado não existe.",
           success: false,
-        })
+        });
       }
-      // adicionar um conexão aqui para ver se o topico existe e o usuario
+
+      // verificar se a postagem já está publicada
       connection.query(
-        "INSERT INTO Post(content, image ,User_idUser, Topic_idTopic) VALUES(?, ?, ?, ?) ",
-        [content, image ,User_idUser, Topic_idTopic],
-        (errInsert, resultInsert) => {
-          if (errInsert) {
+        "SELECT * FROM Post where content = ? AND User_idUser = ?",
+        [content, User_idUser],
+        (errPost, resultPost) => {
+          if (errPost) {
             return res.status(500).json({
-              message: "Erro ao criar uma publicação.",
+              message: "Erro ao verificar post existente.",
               success: false,
-              data: errInsert,
-            })
-          } else {
-            return res.status(201).json({
-              success: true,
-              message: "Publicação criada com sucesso.",
-              data: resultInsert,
-            })
+              data: errPost,
+            });
           }
+
+          if (resultPost.length > 0) {
+            return res.status(400).json({
+              message:
+                "Já existe uma postagem igual. Por favor, tente novamente.",
+              success: false,
+            });
+          }
+          connection.query(
+            "INSERT INTO Post(content, image ,User_idUser, Topic_idTopic) VALUES(?, ?, ?, ?) ",
+            [content, image, User_idUser, Topic_idTopic],
+            (errInsert, resultInsert) => {
+              if (errInsert) {
+                return res.status(500).json({
+                  message: "Erro ao criar uma publicação.",
+                  success: false,
+                  data: errInsert,
+                });
+              }
+              return res.status(201).json({
+                success: true,
+                message: "Publicação criada com sucesso.",
+                data: { id: resultInsert.insertId },
+              });
+            }
+          );
         }
-      )
+      );
     }
-  )
-}
+  );
+};
 
 exports.updateContentPost = (req, res) => {
-  const idPost = req.params.IdPost
-  const User_idUser = req.data.id
-  const { content } = req.body
+  const idPost = req.params.IdPost;
+  const User_idUser = req.data.id;
+  const { content } = req.body;
 
   connection.query(
     "SELECT * FROM Post WHERE idPost = ?",
     [idPost],
-    (err, result) => {
-      if (err) {
+    (errUpdate, resultUpdate) => {
+      if (errUpdate) {
         return res.status(500).json({
           message: "Erro ao se conectar com o servidor.",
           success: false,
-          data: err,
-        })
+          data: errUpdate,
+        });
       }
 
-      if (result.length === 0) {
+      if (resultUpdate.length === 0) {
         return res.status(404).json({
           success: false,
           message: `A postagem não foi encontrada para ser atualizada!`,
-        })
-      } else {
-        connection.query(
-          `UPDATE FROM Post where idPost = ? AND User_idUser = ?`,
-          [content, idPost, User_idUser],
-          (err, result) => {
-            if (err) {
-              return res.status(500).json({
-                message: "Erro ao atualizar o conteudo da Postagem.",
-                success: false,
-                data: err,
-              })
-            } else {
-              return res.status(201).json({
-                message: "Conteúdo da postagem foi atualizado com sucesso!",
-                success: true,
-                data: result,
-              })
-            }
-          }
-        )
+        });
       }
+      connection.query(
+        `UPDATE Post FROM SET content = ? where idPost = ? AND User_idUser = ?`,
+        [content, idPost, User_idUser],
+        (err, result) => {
+          if (err) {
+            return res.status(500).json({
+              message: "Erro ao atualizar o conteudo da Postagem.",
+              success: false,
+              data: err,
+            });
+          }
+          return res.status(201).json({
+            message: "Conteúdo da postagem foi atualizado com sucesso!",
+            success: true,
+            data: result,
+          });
+        }
+      );
     }
-  )
-}
+  );
+};
 
 exports.updateImagePost = (req, res) => {
-  const idPost = req.params.idPost
-  const User_idUser = req.data.id
-  const image = req.file ? req.file.filename : null
+  const idPost = req.params.idPost;
+  const User_idUser = req.data.id;
+  const image = req.file ? req.file.filename : null;
 
   connection.query(
     "SELECT * FROM Post WHERE idPost = ?",
@@ -177,43 +225,41 @@ exports.updateImagePost = (req, res) => {
           message: "Erro ao se conectar com o servidor.",
           success: false,
           data: err,
-        })
+        });
       }
 
       if (result.length === 0) {
         return res.status(404).json({
           success: false,
           message: `A postagem não foi encontrada para ser atualizada!`,
-        })
-      } else {
-        connection.query(
-          `UPDATE Post SET image = ? where idPost = ? AND User_idUser = ?
-          `,
-          [image, idPost, User_idUser],
-          (err, result) => {
-            if (err) {
-              return res.status(500).json({
-                message: "Erro ao atualizar a imagem da Postagem.",
-                success: false,
-                data: err,
-              })
-            } else {
-              return res.status(201).json({
-                message: "A imagem da postagem foi atualizada com sucesso!",
-                success: true,
-                data: result,
-              })
-            }
-          }
-        )
+        });
       }
+      connection.query(
+        `UPDATE Post SET image = ? where idPost = ? AND User_idUser = ?
+          `,
+        [image, idPost, User_idUser],
+        (errUpdate, resultUpdate) => {
+          if (errUpdate) {
+            return res.status(500).json({
+              message: "Erro ao atualizar a imagem da Postagem.",
+              success: false,
+              data: errUpdate,
+            });
+          }
+          return res.status(200).json({
+            message: "A imagem da postagem foi atualizada com sucesso!",
+            success: true,
+            data: resultUpdate,
+          });
+        }
+      );
     }
-  )
-}
+  );
+};
 
 exports.deletePost = (req, res) => {
-  const idPost = req.params.idPost
-  const User_idUser = req.data.id
+  const idPost = req.params.idPost;
+  const User_idUser = req.data.id;
 
   connection.query(
     "SELECT * FROM Post where idPost = ?",
@@ -224,7 +270,7 @@ exports.deletePost = (req, res) => {
           message: "Erro ao se conectar com o servidor.",
           success: false,
           data: err,
-        })
+        });
       }
 
       if (result.length === 0) {
@@ -232,30 +278,34 @@ exports.deletePost = (req, res) => {
           message: `Não existe postagens ainda. Por favor, crie para poder deletar alguma.`,
           success: false,
           data: err,
-        })
-      } else {
-        connection.query(
-          `DELETE FROM Post WHERE idPost = ? AND User_idUser = ?
-          `,
-          [idPost, User_idUser],
-          (err, result) => {
-            if (err) {
-              return res.status(500).json({
-                message: "Erro ao excluir a postagem. Tente novamente.",
-                success: false,
-                data: err,
-              })
-            }
-           else {
-              return res.status(201).json({
-                message: "Exclusão da postagem realizada com sucesso",
-                success: true,
-                data: result,
-              })
-            }
-          }
-        )
+        });
       }
+      if (result[0].User_idUser !== User_idUser) {
+        return res.status(403).json({
+          success: false,
+          message: "Você não tem permissão para excluir esta postagem.",
+        });
+      }
+      connection.query(
+        `DELETE FROM Post WHERE idPost = ? AND User_idUser = ?
+          `,
+        [idPost, User_idUser],
+        (err, result) => {
+          if (err) {
+            return res.status(500).json({
+              message: "Erro ao excluir a postagem. Tente novamente.",
+              success: false,
+              data: err,
+            });
+          } else {
+            return res.status(200).json({
+              message: "Exclusão da postagem realizada com sucesso",
+              success: true,
+              data: result,
+            });
+          }
+        }
+      );
     }
-  )
-}
+  );
+};
